# Api/swagger/swagger.json
{
    "openapi": "3.0.1",
    "info": {
      "title": "Product Management API",
      "description": "API for managing products, users, and reports, including integration with Open Food Facts.",
      "version": "1.0.0"
    },
    "servers": [
      {
        "url": "http://localhost:8080/api",
        "description": "Local server"
      }
    ],
    "paths": {
      "/products": {
        "get": {
          "summary": "Get all products",
          "description": "Retrieve all products from the database. Updates products if they are outdated.",
          "responses": {
            "200": {
              "description": "A list of products",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/Product"
                    }
                  }
                }
              }
            },
            "500": {
              "description": "Server error"
            }
          }
        },
        "post": {
          "summary": "Create a new product",
          "description": "Add a new product to the database. Automatically generates a price if none is provided.",
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductInput"
                }
              }
            }
          },
          "responses": {
            "201": {
              "description": "Product created successfully",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Product"
                  }
                }
              }
            },
            "400": {
              "description": "Product already exists"
            },
            "500": {
              "description": "Server error"
            }
          }
        }
      },
      "/products/{barCode}": {
        "get": {
          "summary": "Fetch product information",
          "description": "Fetch product details from Open Food Facts using the product barcode.",
          "parameters": [
            {
              "name": "barCode",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string"
              },
              "description": "The barcode of the product"
            }
          ],
          "responses": {
            "200": {
              "description": "Product details fetched successfully",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Product"
                  }
                }
              }
            },
            "404": {
              "description": "Product not found"
            },
            "500": {
              "description": "Server error"
            }
          }
        },
        "put": {
          "summary": "Update a product",
          "description": "Update product details using the barcode.",
          "parameters": [
            {
              "name": "barCode",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string"
              },
              "description": "The barcode of the product to update"
            }
          ],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductUpdate"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Product updated successfully",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Product"
                  }
                }
              }
            },
            "404": {
              "description": "Product not found"
            },
            "500": {
              "description": "Server error"
            }
          }
        },
        "delete": {
          "summary": "Delete a product",
          "description": "Delete a product using its barcode.",
          "parameters": [
            {
              "name": "barCode",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string"
              },
              "description": "The barcode of the product to delete"
            }
          ],
          "responses": {
            "200": {
              "description": "Product deleted successfully",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "message": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            },
            "404": {
              "description": "Product not found"
            },
            "500": {
              "description": "Server error"
            }
          }
        }
      },
      "/products/search": {
        "get": {
          "summary": "Search products by name",
          "description": "Search for products using a query string.",
          "parameters": [
            {
              "name": "query",
              "in": "query",
              "required": true,
              "schema": {
                "type": "string"
              },
              "description": "The search query"
            }
          ],
          "responses": {
            "200": {
              "description": "Search results",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/Product"
                    }
                  }
                }
              }
            },
            "404": {
              "description": "No products found"
            },
            "500": {
              "description": "Server error"
            }
          }
        }
      },
      "/reports": {
        "post": {
          "summary": "Generate a report",
          "description": "Generate a report containing sales, revenue, and popular products.",
          "responses": {
            "200": {
              "description": "Report generated successfully",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Report"
                  }
                }
              }
            },
            "500": {
              "description": "Failed to generate report"
            }
          }
        }
      },
      "/users/profile": {
        "get": {
          "summary": "Get user profile",
          "description": "Retrieve the profile of the logged-in user (protected route).",
          "responses": {
            "200": {
              "description": "User profile fetched successfully",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/User"
                  }
                }
              }
            },
            "404": {
              "description": "User not found"
            },
            "500": {
              "description": "Server error"
            }
          }
        },
        "put": {
          "summary": "Update user profile",
          "description": "Update user information (protected route).",
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserUpdate"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "User updated successfully",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/User"
                  }
                }
              }
            },
            "404": {
              "description": "User not found"
            },
            "500": {
              "description": "Server error"
            }
          }
        },
        "delete": {
          "summary": "Delete user account",
          "description": "Delete the logged-in user's account (protected route).",
          "responses": {
            "200": {
              "description": "User deleted successfully",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "message": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            },
            "404": {
              "description": "User not found"
            },
            "500": {
              "description": "Server error"
            }
          }
        }
      }
    },
    "components": {
      "schemas": {
        "Product": {
          "type": "object",
          "properties": {
            "name": { "type": "string" },
            "category": { "type": "string" },
            "ingredients": { "type": "string" },
            "imageUrl": { "type": "string" },
            "barCode": { "type": "string" },
            "price": { "type": "number" },
            "ratings": {
              "type": "array",
              "items": { "type": "number" }
            }
          }
        },
        "ProductInput": {
          "type": "object",
          "properties": {
            "name": { "type": "string" },
            "category": { "type": "string" },
            "ingredients": { "type": "string" },
            "imageUrl": { "type": "string" },
            "barCode": { "type": "string" },
            "price": { "type": "number" }
          },
          "required": ["name", "category", "barCode"]
        },
        "ProductUpdate": {
          "type": "object",
          "properties": {
            "name": { "type": "string" },
            "description": { "type": "string" },
            "price": { "type": "number" },
            "category": { "type": "string" },
            "stock": { "type": "number" },
            "imageUrl": { "type": "string" }
          }
        },
        "Report": {
          "type": "object",
          "properties": {
            "sales": { "type": "integer" },
            "revenue": { "type": "number" },
            "averageOrderPrice": { "type": "number" },
            "mostPurchasedProducts": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "productId": { "type": "string" },
                  "name": { "type": "string" },
                  "purchaseCount": { "type": "integer" }
                }
              }
            }
          }
        },
        "User": {
          "type": "object",
          "properties": {
            "username": { "type": "string" },
            "email": { "type": "string" }
          }
        },
        "UserUpdate": {
          "type": "object",
          "properties": {
            "username": { "type": "string" },
            "email": { "type": "string" }
          }
        }
      }
    }
  }
  